% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{dataset}
\alias{dataset}
\title{Retrieve a bundled dataset}
\usage{
dataset(name)
}
\arguments{
\item{name}{A dataset name. See \link{datasets} for a listing of available
options.}
}
\value{
A bundled dataset as a \link[tibble:tibble-package]{tibble}.
}
\description{
\code{\link[=dataset]{dataset()}} returns one of the datasets bundled with
\link[=dtlg.data-package]{dtlg.data}.
}
\examples{
dataset("adsl_small")

}
