% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{datasets}
\alias{datasets}
\title{dtlg.data datasets}
\format{
A \link[tibble:tibble-package]{tibble} of bundled datasets of 4
variables:

\describe{
\item{\code{name}}{Name of the dataset.}
\item{\code{nrow}}{Number of observations in the dataset.}
\item{\code{ncol}}{Number of variables in the dataset.}
\item{\code{description}}{Brief description of the dataset.}
}
}
\usage{
datasets
}
\description{
\code{datasets} is a table listing datasets bundled with
\link[=dtlg.data-package]{dtlg.data}. Data is generated with the
\link[random.cdisc.data:random.cdisc.data-package]{random.cdisc.data} package.

Use the \code{name} field as input to \code{\link[=dataset]{dataset()}} to retrieve that dataset into
memory.
}
\examples{
datasets

}
\keyword{datasets}
