% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtlg.data-package.R
\docType{package}
\name{dtlg.data-package}
\alias{dtlg.data}
\alias{dtlg.data-package}
\title{dtlg.data: Benchmark Data for the 'dtlg' Package}
\description{
A set of datasets to support benchmarking with the package dtlg.
}
\author{
\strong{Maintainer}: Ramiro Magno \email{ramiro.morgado@ascent.io} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

Authors:
\itemize{
  \item Max Ebenezer-Brown \email{max.ebenezer-brown@ascent.io}
}

Other contributors:
\itemize{
  \item Ascent Digital Services (https://www.ascent.io) [copyright holder, funder]
}

}
\keyword{internal}
